feat: First version
First version of env2conffile script.
Contains support for yaml, json and handlebars renderers ie. basic support for handlebars
templates exists. Output target can be stdout, stderr (just because) or file. User can
specify environment variable prefix and/or suffix which is used to filter the variables.
If no filter is given, then the renderer context is the full process.env. Both filters
can be given in which case both must match for the variable to end up in the context.