fix(Generate): Change the order in which the variables are added to the final env object
Earlier the .env file definitions overwrote those defined on command line or in the environment.
This did not allow overriding the default values and is considered to be against how the software is
supposed to work. The order has now been changed so that the .env file values are considered as the
default values that can be replaced with those from the environment.

fix #3